backend/requirements.txt
djangorestframework
djangorestframework-gis
django-cors-headers
django-phonenumber-field
django-nested-admin
django-admin-rangefilter
git+https://github.com/lukasvinclav/django-admin-numeric-filter.git@c5718f59eea66f67bbaa2ee542206719dfd43df3
phonenumbers
Django
gunicorn
Pillow
psycopg2-binary
sentry-sdk
pytest-django
pytest
python-dotenv
pyyaml
django-debug-toolbar
requests
simplejson
django-solo
django-admin-autocomplete-filter
django-more-admin-filters
python-dateutil
django-filter
django-tinymce
XlsxWriter
drf-spectacular
drf-nested-routers
vokativ
django-bootstrap5
